+++ src/index.ts
@@ -15,6 +15,8 @@ export { AuroraReveal } from './components/reveal'
 export { AuroraCounter } from './components/counter'
 export { AuroraCursor } from './components/cursor'
 export { AuroraNebula } from './components/nebula'
+export { AuroraScramble } from './components/scramble'
+export { AuroraTypewriter } from './components/typewriter'
 
 export { AuroraElement } from './core/base'
 export { register } from './core/register'